requirements.txt
# conda install -c anaconda pyqt
# conda install -c conda-forge opencv
imgviz
natsort
scikit-image
cv2
numpy
pandasy